Updated GPT prompt, feeder fields, and removed recommendation section.
The GPT prompt was reworked in an attempt to prevent the problematic racial stereotyping when generating restaurant recommendations, however, given the raw information we are feeding the model, this was impossible without the recommendation devolving into any generic restaurant. So, the restaurant recommendation section was removed to prevent problematic recommendations. As well, so of the census data fields being used or calculated were the incorrect ones, or fields were available and didn't need to be engineered. The correct census fields have been added. The double census tract issue should have been fixed as well.  Updated known issues sections.